Created Image class to encapsulate some transforms with QImage
