+++ libraries/image/src/image/Image.h
@@ -1,9 +1,10 @@
+#pragma once
 //
 //  Image.h
-//  image/src/image
+//  image/src/Image
 //
-//  Created by Clement Brisset on 4/5/2017.
-//  Copyright 2017 High Fidelity, Inc.
+//  Created by Olivier Prat on 29/3/2019.
+//  Copyright 2019 High Fidelity, Inc.
 //
 //  Distributed under the Apache License, Version 2.0.
 //  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
@@ -12,85 +13,85 @@
 #ifndef hifi_image_Image_h
 #define hifi_image_Image_h
 
-#include <QVariant>
 #include <QImage>
-
-#include <gpu/Texture.h>
 
 #include "ColorChannel.h"
 
-class QByteArray;
+#include <glm/fwd.hpp>
+#include <glm/vec2.hpp>
+#include <GLMHelpers.h>
 
 namespace image {
 
-extern const QImage::Format QIMAGE_HDRFORMAT;
+    class Image {
+    public:
 
-std::function<gpu::uint32(const glm::vec3&)> getHDRPackingFunction();
-std::function<glm::vec3(gpu::uint32)> getHDRUnpackingFunction();
+        enum Format {
+            Format_Invalid = QImage::Format_Invalid,
+            Format_Mono = QImage::Format_Mono,
+            Format_MonoLSB = QImage::Format_MonoLSB,
+            Format_Indexed8 = QImage::Format_Indexed8,
+            Format_RGB32 = QImage::Format_RGB32,
+            Format_ARGB32 = QImage::Format_ARGB32,
+            Format_ARGB32_Premultiplied = QImage::Format_ARGB32_Premultiplied,
+            Format_RGB16 = QImage::Format_RGB16,
+            Format_ARGB8565_Premultiplied = QImage::Format_ARGB8565_Premultiplied,
+            Format_RGB666 = QImage::Format_RGB666,
+            Format_ARGB6666_Premultiplied = QImage::Format_ARGB6666_Premultiplied,
+            Format_RGB555 = QImage::Format_RGB555,
+            Format_ARGB8555_Premultiplied = QImage::Format_ARGB8555_Premultiplied,
+            Format_RGB888 = QImage::Format_RGB888,
+            Format_RGB444 = QImage::Format_RGB444,
+            Format_ARGB4444_Premultiplied = QImage::Format_ARGB4444_Premultiplied,
+            Format_RGBX8888 = QImage::Format_RGBX8888,
+            Format_RGBA8888 = QImage::Format_RGBA8888,
+            Format_RGBA8888_Premultiplied = QImage::Format_RGBA8888_Premultiplied,
+            Format_Grayscale8 = QImage::Format_Grayscale8,
+            Format_R11G11B10F = QImage::Format_RGB30,
+            Format_PACKED_FLOAT = Format_R11G11B10F
+        };
 
-namespace TextureUsage {
+        using AspectRatioMode = Qt::AspectRatioMode;
+        using TransformationMode = Qt::TransformationMode;
 
-enum Type {
-    DEFAULT_TEXTURE,
-    STRICT_TEXTURE,
-    ALBEDO_TEXTURE,
-    NORMAL_TEXTURE,
-    BUMP_TEXTURE,
-    SPECULAR_TEXTURE,
-    METALLIC_TEXTURE = SPECULAR_TEXTURE, // for now spec and metallic texture are the same, converted to grey
-    ROUGHNESS_TEXTURE,
-    GLOSS_TEXTURE,
-    EMISSIVE_TEXTURE,
-    CUBE_TEXTURE,
-    OCCLUSION_TEXTURE,
-    SCATTERING_TEXTURE = OCCLUSION_TEXTURE,
-    LIGHTMAP_TEXTURE,
-    UNUSED_TEXTURE
-};
+        Image() {}
+        Image(int width, int height, Format format) : _data(width, height, (QImage::Format)format) {}
+        Image(const QImage& data) : _data(data) {}
+        void operator=(const QImage& image) {
+            _data = image;
+        }
 
-using TextureLoader = std::function<gpu::TexturePointer(QImage&&, const std::string&, bool, gpu::BackendTarget, const std::atomic<bool>&)>;
-TextureLoader getTextureLoaderForType(Type type, const QVariantMap& options = QVariantMap());
+        bool isNull() const { return _data.isNull(); }
 
-gpu::TexturePointer create2DTextureFromImage(QImage&& image, const std::string& srcImageName,
-                                             bool compress, gpu::BackendTarget target, const std::atomic<bool>& abortProcessing);
-gpu::TexturePointer createStrict2DTextureFromImage(QImage&& image, const std::string& srcImageName,
-                                                   bool compress, gpu::BackendTarget target, const std::atomic<bool>& abortProcessing);
-gpu::TexturePointer createAlbedoTextureFromImage(QImage&& image, const std::string& srcImageName,
-                                                 bool compress, gpu::BackendTarget target, const std::atomic<bool>& abortProcessing);
-gpu::TexturePointer createEmissiveTextureFromImage(QImage&& image, const std::string& srcImageName,
-                                                   bool compress, gpu::BackendTarget target, const std::atomic<bool>& abortProcessing);
-gpu::TexturePointer createNormalTextureFromNormalImage(QImage&& image, const std::string& srcImageName,
-                                                       bool compress, gpu::BackendTarget target, const std::atomic<bool>& abortProcessing);
-gpu::TexturePointer createNormalTextureFromBumpImage(QImage&& image, const std::string& srcImageName,
-                                                     bool compress, gpu::BackendTarget target, const std::atomic<bool>& abortProcessing);
-gpu::TexturePointer createRoughnessTextureFromImage(QImage&& image, const std::string& srcImageName,
-                                                    bool compress, gpu::BackendTarget target, const std::atomic<bool>& abortProcessing);
-gpu::TexturePointer createRoughnessTextureFromGlossImage(QImage&& image, const std::string& srcImageName,
-                                                         bool compress, gpu::BackendTarget target, const std::atomic<bool>& abortProcessing);
-gpu::TexturePointer createMetallicTextureFromImage(QImage&& image, const std::string& srcImageName,
-                                                   bool compress, gpu::BackendTarget target, const std::atomic<bool>& abortProcessing);
-gpu::TexturePointer createCubeTextureFromImage(QImage&& image, const std::string& srcImageName,
-                                               bool compress, gpu::BackendTarget target, const std::atomic<bool>& abortProcessing);
-gpu::TexturePointer createCubeTextureFromImageWithoutIrradiance(QImage&& image, const std::string& srcImageName,
-                                                                bool compress, gpu::BackendTarget target, const std::atomic<bool>& abortProcessing);
-gpu::TexturePointer createLightmapTextureFromImage(QImage&& image, const std::string& srcImageName,
-                                                   bool compress, gpu::BackendTarget target, const std::atomic<bool>& abortProcessing); 
-gpu::TexturePointer process2DTextureColorFromImage(QImage&& srcImage, const std::string& srcImageName, bool compress,
-                                                   gpu::BackendTarget target, bool isStrict, const std::atomic<bool>& abortProcessing);
-gpu::TexturePointer process2DTextureNormalMapFromImage(QImage&& srcImage, const std::string& srcImageName, bool compress,
-                                                       gpu::BackendTarget target, bool isBumpMap, const std::atomic<bool>& abortProcessing);
-gpu::TexturePointer process2DTextureGrayscaleFromImage(QImage&& srcImage, const std::string& srcImageName, bool compress,
-                                                       gpu::BackendTarget target, bool isInvertedPixels, const std::atomic<bool>& abortProcessing);
-gpu::TexturePointer processCubeTextureColorFromImage(QImage&& srcImage, const std::string& srcImageName, bool compress,
-                                                     gpu::BackendTarget target, bool generateIrradiance, const std::atomic<bool>& abortProcessing);
+        Format getFormat() const { return (Format)_data.format(); }
+        bool hasAlphaChannel() const { return _data.hasAlphaChannel(); }
 
-} // namespace TextureUsage
+        glm::uint32 getWidth() const { return (glm::uint32)_data.width(); }
+        glm::uint32 getHeight() const { return (glm::uint32)_data.height(); }
+        glm::uvec2 getSize() const { return toGlm(_data.size()); }
+        size_t getByteCount() const { return _data.byteCount(); }
 
-const QStringList getSupportedFormats();
+        QRgb getPixel(int x, int y) const { return _data.pixel(x, y); }
+        void setPixel(int x, int y, QRgb value) {
+            _data.setPixel(x, y, value);
+        }
 
-gpu::TexturePointer processImage(std::shared_ptr<QIODevice> content, const std::string& url, ColorChannel sourceChannel,
-                                 int maxNumPixels, TextureUsage::Type textureType,
-                                 bool compress, gpu::BackendTarget target, const std::atomic<bool>& abortProcessing = false);
+        glm::uint8* editScanLine(int y) { return _data.scanLine(y); }
+        const glm::uint8* getScanLine(int y) const { return _data.scanLine(y); }
+        const glm::uint8* getBits() const { return _data.constBits(); }
+
+        Image getScaled(glm::uvec2 newSize, AspectRatioMode ratioMode, TransformationMode transformationMode = Qt::SmoothTransformation) const;
+        Image getConvertedToFormat(Format newFormat) const;
+        Image getSubImage(QRect rect) const;
+        Image getMirrored(bool horizontal, bool vertical) const;
+
+        // Inplace transformations
+        void invertPixels();
+
+    private:
+
+        QImage _data;
+    };
 
 } // namespace image
